Add havak account

+++ src/main/java/rocks/zipcode/atm/bank/HavakAccount.java
@@ -11,7 +11,7 @@ public class HavakAccount extends Account {
     }
 
     @Override
-    protected boolean canWithdraw(float amount) {
+    protected boolean canWithdraw(int amount) {
         return getBalance() + OVERDRAFT_LIMIT >= amount;
     }
